+++ app/src/main.jsx
@@ -1,9 +1,9 @@
-import * as React from "react";
-import * as ReactDOM from "react-dom";
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
 import '@cybercongress/ui/lib/styles.css';
 import Root from './root';
 
 ReactDOM.render(
-    <Root/>
-    , document.getElementById("root")
+    <Root />,
+    document.getElementById('root'),
 );
