+++ app/src/root.jsx
@@ -1,13 +1,13 @@
-import React from "react";
-import { HashRouter as Router, Route } from "react-router-dom";
-import App from './containers/App/App';
+import React from 'react';
+import { HashRouter as Router, Route } from 'react-router-dom';
 import Stake from './containers/Stake/Stake';
+import Search from './containers/Search';
 
 const Root = () => (
     <Router>
         <div>
-            <Route exact path="/" component={ App } />
-            <Route path="/stake" component={ Stake } />
+            <Route exact path='/' component={ Search } />
+            <Route path='/stake' component={ Stake } />
         </div>
     </Router>
 );
