+++ package.json
@@ -28,6 +28,7 @@
     "react": "15.6.1",
     "react-dom": "15.6.1",
     "react-router-dom": "4.3.1",
+    "unstated": "2.1.1",
     "eslint": "5.14.1",
     "eslint-config-airbnb": "17.1.0",
     "eslint-config-standard-react": "7.0.2",
